find course by id with its categories and calculate average rating for that course

+++ src/main/java/com/projects/spring/udemy/course/dto/CourseInMenu.java
@@ -23,7 +23,7 @@ public class CourseInMenu {
         return averageRating;
     }
 
-    public void setAverageRating(int averageRating) {
+    public void setAverageRating(double averageRating) {
         this.averageRating = averageRating;
     }
